feat(pending changes): deepen the stack of changes (#15070)
Currently, there are multiple situations under which translating a unit requires that
the previously pending changes to the unit be flushed to the repository. One such example
is two different users editing the same unit, and the second user's changes cannot be saved
till the first user's changes to the unit have been committed. This is a performance issue.

To fix this issue, introduce a new PendingUnitChange model that keeps track of the pending
changes to be flushed to the repository. When a unit is translated or saved otherwise, a
PendingUnitChange record is created for the change instead of flushing the changes to the
repository.

The logic to group changes by author has changed as the previous code depended on the changes
to a unit by a different author to have been already committed. The new logic handles these
cases of changes to the same unit by multiple authors correctly.

* address partial changes from code review

* revert inadvertent quotes change

* remove unneeded save method

* remove unneeded import_user_memory

* add has_pending_changes property, squash model migrations, misc fixes

* Add data migration for boolean pending to PendingUnitChange model

* fix data migration for tests

* handle pending units in backups, fix commit grouping and remaining tests, squash migrations.

* allow multiple changes per unit-author in pending unit chang

also, use default for timestamp to allow timestamp restore from backups.

* do not lookup units from database in backup restoration

fix unit.pending -> PendingUnitChange migration script

* delete pending units only if change is applied successfully

* update deletion of pending changes

* load anonymous user lazily in migration

* try adding migration test for pending unit changes

* Fix semver condition for migration testing

* Remove unused code

Co-authored-by: Michal Čihař <michal@cihar.com>