chore(pending change): make retries smarter
Currently, all failing pending changes are retried on every commit, and that happens
around every hour or so. This creates lots of noise in the unit's changelog and
sentry as failing changes are unlikely to resolve themselves. Hence, add a
metadata field to PendingUnitChange to store the translation revision at the time
of the failure, the time of the failure and the weblate version. Failing changes
are only retried when either the revision or weblate version changes or a week
has passed.

Also, failures are of two types - blocking due to missing unit in the translation files
and non-blocking due to issues possibly in TranslateToolkit. Whether a change is
blocking or not is stored in the metadata field as well. If a failed pending change
of blocking type exists, all newer changes to the unit are excluded from selection.
A non-blocking failed pending change might itself not be selected for retrying if
conditions are not met but it does not prevent newer changes to the unit from
selection. If any such changes are successfully applied, the earlier failed changes
are deleted from the table.