feat: push filter to database in commit pending detection (#16102)
* feat: push filter to database in commit pending detection

Push the age cutoff filter to the database when detecting which components might
be eligible for a commit.

* Add MySQLTimestampAdd function to handle interval conversion

HOUR literal is quoted if passed in a `Value` and mariadb/mysql doesn't
accept that.

* add types

* Type anotation fixes

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update weblate/trans/functions.py

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Apply suggestions from code review

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update weblate/trans/functions.py

---------

Co-authored-by: Michal Čihař <michal@cihar.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>