fix(pending change): write correct state to vcs
The state written to VCS should be taken from the PendingUnitChange and not
the unit itself as it might have changed since then. This is especially problematic
when using a commit policy like approved only, and a unit has approved changes but
also subsequent changes waiting for review. In such cases, the approved target string
will be written to the file but the state would be translated and not approved in
absence of this fix.