fix(memory): exclude blank context from matching
The histrical translation memory entries do not store context and are
now suddenly 95% matches instead of 100% ones. This is confusing to
users as nothing has changed and the automatic translation with 100
qualtity threshold no longer works.

The downside of this is that we ignore the actual blank context when
translating strings with a non-blank context.