feat(vcs): extended filtering of allowed VCS URLs
This gives users tigher control on what can be cloned into Weblate. We
now block unsafe protocols such as git or http. The file protocol
blocking in Git has been replaced by this new setting as well.

This also makes it harder to use remote helpers, so the documentation
for it is removed. I don't think anybody is using that anyway, so stop
pretend that this is something we care about.