feat: detect automatically translated units from XLIFF state-qualifier
Add support for detecting machine-translated units in XLIFF files based
on the state-qualifier attribute. Units with "leveraged-mt" or
"mt-suggestion" state qualifiers are now marked as automatically
translated when imported. To implement this, a `is_automatically_translated`
method has been added to the base TranslationUnit class and it is
used to set the `automatically_translated` flag in database during
unit import.