feat: detect automatically translated units from XLIFF state-qualifier (#17294)
* feat: detect automatically translated units from XLIFF state-qualifier

Add support for detecting machine-translated units in XLIFF files based
on the state-qualifier attribute. Units with "leveraged-mt" or
"mt-suggestion" state qualifiers are now marked as automatically
translated when imported. To implement this, a `is_automatically_translated`
method has been added to the base TranslationUnit class and it is
used to set the `automatically_translated` flag in database during
unit import.

* Write automatically_translated status to xliff files

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Fix migration dependency

* fix test

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Michal Čihař <michal@cihar.com>